+++ src/main/java/com/qa/choonz/service/PlaylistService.java
@@ -52,6 +52,9 @@ public class PlaylistService {
     }
 
     public boolean delete(long id) {
+        if (!this.repo.existsById(id)) {
+            throw new PlaylistNotFoundException();
+        }
         this.repo.deleteById(id);
         return !this.repo.existsById(id);
     }
