+++ src/main/java/com/qa/choonz/service/TrackService.java
@@ -53,6 +53,9 @@ public class TrackService {
     }
 
     public boolean delete(long id) {
+        if (!this.repo.existsById(id)) {
+            throw new TrackNotFoundException();
+        }
         this.repo.deleteById(id);
         return !this.repo.existsById(id);
     }
